Reorganize code
