+++ src/lib.rs
@@ -83,12 +83,14 @@ mod api;
 mod connection;
 mod crypto;
 mod lookup;
+mod types;
 pub mod errors;
 
-pub use api::{ApiBuilder, E2eApi, SimpleApi, RecipientKey};
-pub use connection::{Recipient, BlobId};
-pub use crypto::{EncryptedMessage, MessageType};
+pub use api::{ApiBuilder, E2eApi, SimpleApi};
+pub use connection::{Recipient};
+pub use crypto::{EncryptedMessage, RecipientKey};
 pub use lookup::{LookupCriterion, Capabilities};
+pub use types::{MessageType, BlobId};
 
 const MSGAPI_URL: &'static str = "https://msgapi.threema.ch";
