+++ src/types.rs
@@ -1,0 +1,78 @@
+use std::fmt;
+
+use data_encoding::{HEXLOWER, HEXLOWER_PERMISSIVE};
+
+use ::errors::ApiError;
+
+
+/// A message type.
+pub enum MessageType {
+    Text,
+    Image,
+    Video,
+    File,
+    DeliveryReceipt,
+}
+
+impl Into<u8> for MessageType {
+    fn into(self) -> u8 {
+        match self {
+            MessageType::Text => 0x01,
+            MessageType::Image => 0x02,
+            MessageType::Video => 0x13,
+            MessageType::File => 0x17,
+            MessageType::DeliveryReceipt => 0x80,
+        }
+    }
+}
+
+/// A blob ID. Must contain exactly 16 lowercase hexadecimal characters.
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct BlobId(pub [u8; 16]);
+
+impl BlobId {
+    /// Create a new BlobId.
+    pub fn new(id: [u8; 16]) -> Self {
+        BlobId(id)
+    }
+
+    /// Create a new BlobId from a 32 character hexadecimal String.
+    pub fn from_str(id: &str) -> Result<Self, ApiError> {
+        let bytes = HEXLOWER_PERMISSIVE.decode(id.as_bytes()).map_err(|_| ApiError::BadBlobId)?;
+        if bytes.len() != 16 {
+            return Err(ApiError::BadBlobId);
+        }
+        let mut arr = [0; 16];
+        for i in 0..bytes.len() {
+            arr[i] = bytes[i];
+        }
+        Ok(BlobId(arr))
+    }
+}
+
+impl fmt::Display for BlobId {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", HEXLOWER.encode(&self.0))
+    }
+}
+
+
+#[cfg(test)]
+mod test {
+    use super::{BlobId};
+
+    #[test]
+    fn test_blob_id_from_str() {
+        assert!(BlobId::from_str("0123456789abcdef0123456789abcdef").is_ok());
+        assert!(BlobId::from_str("0123456789abcdef0123456789abcdeF").is_ok());
+        assert!(BlobId::from_str("0123456789abcdef0123456789abcde").is_err());
+        assert!(BlobId::from_str("0123456789abcdef0123456789abcdef\n").is_err());
+        assert!(BlobId::from_str("0123456789abcdef0123456789abcdeg").is_err());
+
+        assert_eq!(
+            BlobId::from_str("000102030405060708090a0b0c0d0eff").unwrap(),
+            BlobId::new([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0xa, 0xb, 0xc, 0xd, 0xe, 0xff])
+        );
+    }
+
+}
